requirements.txt
keyboard==0.13.5
pillow==11.3.0
psutil==7.0.0
pystray==0.19.5
pywin32==311
six==1.17.0
